requirements.txt
fastapi
slowapi
redis
python-dotenv
pydantic
celery
requests
google-generativeai
